djangoestudent/djangoestudent/requirements.txt
django
djangorestframework
django-filter
psycopg2-binary==2.9.1
django-cors-headers
markdown